requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=2.1.0=pyhd8ed1ab_0
aiohttp=3.9.3=py311he705e18_0
aiosignal=1.3.1=pyhd8ed1ab_0
anyio=4.3.0=pyhd8ed1ab_0
appnope=0.1.4=pyhd8ed1ab_0
argon2-cffi=23.1.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py311h2725bcf_4
arrow=1.3.0=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pyhd8ed1ab_0
async-lru=2.0.4=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
babel=2.14.0=pyhd8ed1ab_0
beautifulsoup4=4.12.3=pyha770c72_0
bleach=6.1.0=pyhd8ed1ab_0
blinker=1.7.0=pyhd8ed1ab_0
brotli=1.1.0=h0dc2134_1
brotli-bin=1.1.0=h0dc2134_1
brotli-python=1.1.0=py311hdf8f085_1
bzip2=1.0.8=h10d778d_5
c-ares=1.27.0=h10d778d_0
ca-certificates=2024.2.2=h8857fd0_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.3=pyhd8ed1ab_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.16.0=py311hc0b63fd_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
comm=0.2.1=pyhd8ed1ab_0
contourpy=1.2.0=py311h7bea37d_0
cryptography=42.0.5=py311hda9c3b6_0
cycler=0.12.1=pyhd8ed1ab_0
debugpy=1.8.1=py311hdd0406b_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
flatbuffers=23.5.26=he965462_1
fonttools=4.49.0=py311he705e18_0
fqdn=1.5.1=pyhd8ed1ab_0
freetype=2.12.1=h60636b9_2
frozenlist=1.4.1=py311he705e18_0
gast=0.5.4=pyhd8ed1ab_0
giflib=5.2.1=hb7f2c08_3
google-auth=2.28.1=pyhca7485f_0
google-auth-oauthlib=1.2.0=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.59.3=py311h78ff076_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
h5py=3.10.0=nompi_py311hf7b785c_101
hdf5=1.14.3=nompi_h691f4bf_100
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.4=pyhd8ed1ab_0
httpx=0.27.0=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=73.2=hf5e326d_0
idna=3.6=pyhd8ed1ab_0
importlib-metadata=7.0.1=pyha770c72_0
importlib_metadata=7.0.1=hd8ed1ab_0
importlib_resources=6.1.2=pyhd8ed1ab_0
ipykernel=6.29.3=pyh3cd1d5f_0
ipython=8.22.1=pyh707e725_0
ipywidgets=8.1.2=pyhd8ed1ab_0
isoduration=20.11.0=pyhd8ed1ab_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pyhd8ed1ab_0
joblib=1.3.2=pyhd8ed1ab_0
json5=0.9.18=pyhd8ed1ab_0
jsonpointer=2.4=py311h6eed73b_3
jsonschema=4.21.1=pyhd8ed1ab_0
jsonschema-specifications=2023.12.1=pyhd8ed1ab_0
jsonschema-with-format-nongpl=4.21.1=pyhd8ed1ab_0
jupyter=1.0.0=pyhd8ed1ab_10
jupyter-lsp=2.2.3=pyhd8ed1ab_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_console=6.6.3=pyhd8ed1ab_0
jupyter_core=5.7.1=py311h6eed73b_0
jupyter_events=0.9.0=pyhd8ed1ab_0
jupyter_server=2.12.5=pyhd8ed1ab_0
jupyter_server_terminals=0.5.2=pyhd8ed1ab_0
jupyterlab=4.1.2=pyhd8ed1ab_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_1
jupyterlab_server=2.25.3=pyhd8ed1ab_0
jupyterlab_widgets=3.0.10=pyhd8ed1ab_0
keras=2.15.0=pyhd8ed1ab_0
kiwisolver=1.4.5=py311h5fe6e05_1
krb5=1.21.2=hb884880_0
lcms2=2.16=ha2f27b4_0
lerc=4.0.0=hb486fe8_0
libabseil=20230802.1=cxx17_h048a20a_0
libaec=1.1.2=he965462_1
libblas=3.9.0=21_osx64_openblas
libbrotlicommon=1.1.0=h0dc2134_1
libbrotlidec=1.1.0=h0dc2134_1
libbrotlienc=1.1.0=h0dc2134_1
libcblas=3.9.0=21_osx64_openblas
libcurl=8.5.0=h726d00d_0
libcxx=16.0.6=hd57cbcb_0
libdeflate=1.19=ha4e1b8e_0
libedit=3.1.20191231=h0678c8f_2
libev=4.33=h10d778d_2
libexpat=2.5.0=hf0c8a7f_1
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=13_2_0_h97931a8_3
libgfortran5=13.2.0=h2873a65_3
libgrpc=1.59.3=ha7f534c_0
libjpeg-turbo=3.0.0=h0dc2134_1
liblapack=3.9.0=21_osx64_openblas
libnghttp2=1.58.0=h64cf6d3_1
libopenblas=0.3.26=openmp_hfef2a42_0
libpng=1.6.43=h92b6c6a_0
libprotobuf=4.24.4=hc4f2305_0
libre2-11=2023.09.01=h4694dbf_1
libsodium=1.0.18=hbcb3906_1
libsqlite=3.45.1=h92b6c6a_0
libssh2=1.11.0=hd019ec5_0
libtiff=4.6.0=h684deea_2
libwebp-base=1.3.2=h0dc2134_0
libxcb=1.15=hb7f2c08_0
libzlib=1.2.13=h8a1eda9_5
llvm-openmp=17.0.6=hb6ac08f_0
markdown=3.5.2=pyhd8ed1ab_0
markupsafe=2.1.5=py311he705e18_0
matplotlib=3.8.3=py311h6eed73b_0
matplotlib-base=3.8.3=py311h6ff1f5f_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=3.0.2=pyhd8ed1ab_0
ml_dtypes=0.2.0=py311hab14417_2
multidict=6.0.5=py311h5547dcb_0
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.8.0=pyhd8ed1ab_0
nbconvert=7.16.1=pyhd8ed1ab_0
nbconvert-core=7.16.1=pyhd8ed1ab_0
nbconvert-pandoc=7.16.1=pyhd8ed1ab_0
nbformat=5.9.2=pyhd8ed1ab_0
ncurses=6.4=h93d8f39_2
nest-asyncio=1.6.0=pyhd8ed1ab_0
notebook=7.1.1=pyhd8ed1ab_0
notebook-shim=0.2.4=pyhd8ed1ab_0
numpy=1.26.4=py311hc43a94b_0
oauthlib=3.2.2=pyhd8ed1ab_0
openjpeg=2.5.2=h7310d3a_0
openssl=3.2.1=hd75f5a5_0
opt_einsum=3.3.0=pyhc1e730c_2
overrides=7.7.0=pyhd8ed1ab_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.2.1=py311h8f6166a_0
pandoc=3.1.12.2=h694c41f_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=py311hea5c87a_0
pip=24.0=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=4.2.0=pyhd8ed1ab_0
prometheus_client=0.20.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
protobuf=4.24.4=py311h021eaf5_0
psutil=5.9.8=py311he705e18_0
pthread-stubs=0.4=hc929b4f_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.1=pyhd8ed1ab_0
pyasn1-modules=0.3.0=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pyjwt=2.8.0=pyhd8ed1ab_1
pyobjc-core=10.1=py311h9b70068_0
pyobjc-framework-cocoa=10.1=py311h9b70068_0
pyopenssl=24.0.0=pyhd8ed1ab_0
pyparsing=3.1.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.8=h9f0c242_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-fastjsonschema=2.19.1=pyhd8ed1ab_0
python-flatbuffers=23.5.26=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.11=4_cp311
pytz=2024.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0.1=py311h2725bcf_1
pyzmq=25.1.2=py311h889d6d6_0
qtconsole-base=5.5.1=pyha770c72_0
qtpy=2.4.1=pyhd8ed1ab_0
re2=2023.09.01=hb168e87_1
readline=8.2=h9e318b2_1
referencing=0.33.0=pyhd8ed1ab_0
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rpds-py=0.18.0=py311hd64b9fd_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.4.1.post1=py311he2b4599_0
scipy=1.12.0=py311h86d0cd9_2
send2trash=1.8.2=pyhd1c38e8_0
setuptools=69.1.1=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h225ccf5_0
sniffio=1.3.1=pyhd8ed1ab_0
soupsieve=2.5=pyhd8ed1ab_1
stack_data=0.6.2=pyhd8ed1ab_0
tensorboard=2.15.2=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py311h892b619_1
tensorflow=2.15.0=cpu_py311h7a9ae38_2
tensorflow-base=2.15.0=cpu_py311hbca55e7_2
tensorflow-estimator=2.15.0=cpu_py311hd6d7b2b_2
termcolor=2.4.0=pyhd8ed1ab_0
terminado=0.18.0=pyh31c8845_0
threadpoolctl=3.3.0=pyhc1e730c_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.13=h1abcd95_1
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.4=py311he705e18_0
traitlets=5.14.1=pyhd8ed1ab_0
types-python-dateutil=2.8.19.20240106=pyhd8ed1ab_0
typing-extensions=4.10.0=hd8ed1ab_0
typing_extensions=4.10.0=pyha770c72_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2024a=h0c530f3_0
uri-template=1.3.0=pyhd8ed1ab_0
urllib3=2.2.1=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=1.13=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=1.7.0=pyhd8ed1ab_0
werkzeug=3.0.1=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
widgetsnbextension=4.0.10=pyhd8ed1ab_0
wrapt=1.14.1=py311h5547dcb_1
xorg-libxau=1.0.11=h0dc2134_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.6=h775f41a_0
yaml=0.2.5=h0d85af4_2
yarl=1.9.4=py311he705e18_0
zeromq=4.3.5=h93d8f39_0
zipp=3.17.0=pyhd8ed1ab_0
zstd=1.5.5=h829000d_0
